refactor: split trade_screen Market tab into part file under size cap (Refs #3594)
Extracts `_MarketTabContent` and `_SectionedTradeableCommodities` from
`trade_screen.dart` into the new `trade_screen_market_tab.dart` part
fragment, continuing the file's established part-based decomposition
(`trade_screen_deal_book.dart`, `trade_screen_market_row.dart`).

Behavior-preserving move: no logic, keys, or widget structure changed.
Brings `trade_screen.dart` from 1381 to 614 physical lines, satisfying
the issue #3594 AC6 (<=700 lines) and the code-review >1000-physical-line
split guidance.
